frontend: Allow TopBar to center the cluster chooser button
The cluster chooser is supposed to be centered in the top bar, which
means the top bar needs to change its left margin according to whether
the sidebar is open or not.

The code that checks whether the sidebar is open (in order to
calculate the topbar size) was being tied to whether the screen is
small, but it shouldn't have to take that into account.